quiet the compiler

+++ TRIOS-MOBILE/FirstViewController.m
@@ -171,7 +171,7 @@
 {
 }
 
--(int)tableView:tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger)tableView:tableView numberOfRowsInSection:(NSInteger)section
 {
    return [status.names count];
 }
